Update page metadata
fixes #23

It seems like Twitter might require a @name to display, but I don't think it'll let me validate a local site. Slack seems to work currently, but even Google+ and Facebook are throwing a fit. I think adding the description field should help and I have a feeling that they might want an absolute URL for the image. AFAICT, this is a bit of "guess and check" though.

It sounds like flo doesn't want a twitter account, so I'm wonder if for the purpose of twitter cards we use someone else's account? Whether that be GNOME or elementary or a contributor?